Fix lint issues with theme provider
