+++ frontend/src/contexts/ThemeContextDefinition.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+export interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
