+++ frontend/src/pages/GodMode.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import Navbar from "../components/Navbar";
-import { useTheme } from "../contexts/ThemeContext";
+import { useTheme } from "../contexts/useTheme";
 import { useNavigate } from "react-router-dom";
 import {
   getCurrentPrivileges,
