+++ frontend/src/pages/SOSTools.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import Navbar from "../components/Navbar";
-import { useTheme } from "../contexts/ThemeContext";
+import { useTheme } from "../contexts/useTheme";
 import "./Chatbot.css"; 
 import "./SOSTools.css";
